Move network constants to shared folder
