+++ app/scripts/controllers/detect-tokens.js
@@ -2,7 +2,7 @@ import Web3 from 'web3'
 import contracts from '@metamask/contract-metadata'
 import { warn } from 'loglevel'
 import SINGLE_CALL_BALANCES_ABI from 'single-call-balance-checker-abi'
-import { MAINNET_CHAIN_ID } from './network/enums'
+import { MAINNET_CHAIN_ID } from '../../../shared/constants/network'
 
 // By default, poll every 3 minutes
 const DEFAULT_INTERVAL = 180 * 1000
