+++ app/scripts/controllers/preferences.js
@@ -7,7 +7,7 @@ import ethers from 'ethers'
 import log from 'loglevel'
 import { isPrefixedFormattedHexString } from '../lib/util'
 import { LISTED_CONTRACT_ADDRESSES } from '../../../shared/constants/tokens'
-import { NETWORK_TYPE_TO_ID_MAP } from './network/enums'
+import { NETWORK_TYPE_TO_ID_MAP } from '../../../shared/constants/network'
 
 export default class PreferencesController {
   /**
